requirements.txt
gradio==3.50.2
tensorflow==2.12.0
numpy
Pillow
